match some more math

+++ include/math.h
@@ -8,6 +8,6 @@ u32 AbsMax(s32 a, s32 b);
 u32 sub_8085530(s32 a, s32 b);
 s16 sub_808558C(u16 a, u16 b, u8 c);
 u32 sub_80855C0(s32 a, s32 b, s32 c, u8 d);
-s16 sub_8085654(s32, s32, s32, u8, u8);
+s32 sub_8085654(s32, s32, s32, u8, u8);
 
 #endif // GUARD_MATH_H
